add suspense for vega charts
